requirements.txt
shiny
ultralytics
numpy
Pillow
opencv-python-headless
nest-asyncio